Player/slime interaction, pause 'menu'
Player knocked back and loses health if it touches the slime
-Player move diagonally based on where the slime is when they collide
-Jumpy(just teleports back, rather than 'sliding' back)
Press 'p' to pause the game, press 'esc' to resume
NOTE-- I started this with the Will2.py file, then added the stuff above. The IceNova only seems to work only on one of the two slimes. As far as I know this was the case before I modified the code, as I didn't mess with the slime or IceNova's hitbox or interaction at all.